requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.4.0=pypi_0
accelerate=0.17.1=pypi_0
aiofiles=23.1.0=pypi_0
aiohttp=3.8.1=py39hb9d737c_1
aiosignal=1.3.1=pyhd8ed1ab_0
altair=4.2.2=pypi_0
anyio=3.6.2=pypi_0
appdirs=1.4.4=pyh9f0ad1d_0
async-timeout=4.0.2=pyhd8ed1ab_0
attrs=22.2.0=pyh71513ae_0
bitsandbytes=0.37.2=pypi_0
blas=1.0=mkl
blessed=1.20.0=pypi_0
bottleneck=1.3.5=py39h7deecbd_0
brotlipy=0.7.0=py39h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
ca-certificates=2022.12.7=ha878542_0
cachetools=5.3.0=pypi_0
certifi=2022.12.7=pyhd8ed1ab_0
cffi=1.15.1=py39h5eee18b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.3=unix_pyhd8ed1ab_2
colorama=0.4.6=pyhd8ed1ab_0
cpm-kernels=1.0.11=pypi_0
cryptography=39.0.1=py39h9ce1e76_0
cuda-cudart=11.7.99=0
cuda-cupti=11.7.101=0
cuda-libraries=11.7.1=0
cuda-nvrtc=11.7.99=0
cuda-nvtx=11.7.91=0
cuda-runtime=11.7.1=0
cycler=0.11.0=pyhd8ed1ab_0
datasets=2.10.1=pypi_0
deepspeed=0.8.3=pypi_0
dill=0.3.6=pypi_0
docker-pycreds=0.4.0=py_0
entrypoints=0.4=pypi_0
et_xmlfile=1.1.0=pyhd8ed1ab_0
fastapi=0.95.0=pypi_0
ffmpeg=4.3=hf484d3e_0
ffmpy=0.3.0=pypi_0
filelock=3.9.0=py39h06a4308_0
fire=0.5.0=pypi_0
flit-core=3.8.0=py39h06a4308_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.3.3=py39h5eee18b_0
fsspec=2023.3.0=pypi_0
giflib=5.2.1=h5eee18b_3
gitdb=4.0.10=pyhd8ed1ab_0
gitpython=3.1.31=pyhd8ed1ab_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py39heeb90bb_0
gnutls=3.6.15=he1e5248_0
google-auth=2.17.2=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
gpustat=1.1=pypi_0
gradio=3.23.0=pypi_0
grpcio=1.53.0=pypi_0
h11=0.14.0=pypi_0
hjson=3.1.0=pypi_0
httpcore=0.16.3=pypi_0
httpx=0.23.3=pypi_0
huggingface-hub=0.13.3=pypi_0
icetk=0.0.4=pypi_0
idna=3.4=py39h06a4308_0
importlib-metadata=6.1.0=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
jinja2=3.1.2=py39h06a4308_0
joblib=1.2.0=pyhd8ed1ab_0
jpeg=9e=h5eee18b_1
jsonschema=4.17.3=pypi_0
kiwisolver=1.4.4=py39h6a678d5_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libcublas=11.10.3.66=0
libcufft=10.7.2.124=h4fbf590_0
libcufile=1.6.0.25=0
libcurand=10.3.2.56=0
libcusolver=11.4.0.1=0
libcusparse=11.7.4.91=0
libdeflate=1.17=h5eee18b_0
libffi=3.4.2=h6a678d5_6
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=12.2.0=h69a702a_19
libgfortran5=12.2.0=h337968e_19
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.2=h7f8727e_0
libnpp=11.7.4.75=0
libnvjpeg=11.8.0.2=0
libpng=1.6.39=h5eee18b_0
libprotobuf=3.20.3=he621ea3_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.16.0=h27cfd23_0
libtiff=4.5.0=h6a678d5_2
libunistring=0.9.10=h27cfd23_0
libwebp=1.2.4=h11a3e52_1
libwebp-base=1.2.4=h5eee18b_1
linkify-it-py=2.0.0=pypi_0
loralib=0.1.1=pypi_0
lz4-c=1.9.4=h6a678d5_0
markdown=3.4.3=pypi_0
markdown-it-py=2.2.0=pypi_0
markupsafe=2.1.1=py39h7f8727e_0
matplotlib-base=3.4.3=py39h2fa2bec_2
mdit-py-plugins=0.3.3=pypi_0
mdurl=0.1.2=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py39h7f8727e_0
mkl_fft=1.3.1=py39hd3c417c_0
mkl_random=1.2.2=py39h51133e4_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.2.1=py39h06a4308_0
multidict=6.0.2=py39h5eee18b_0
multiprocess=0.70.14=pypi_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=2.8.4=py39h06a4308_1
ninja=1.11.1=pypi_0
nltk=3.8.1=pypi_0
numexpr=2.8.4=py39he184ba9_0
numpy=1.23.5=py39h14f4228_0
numpy-base=1.23.5=py39h31eccc5_0
nvidia-ml-py=11.525.84=pypi_0
oauthlib=3.2.2=pypi_0
openai=0.27.2=pyhd8ed1ab_0
openh264=2.1.1=h4ff587b_0
openpyxl=3.0.9=pyhd8ed1ab_0
openssl=1.1.1t=h7f8727e_0
orjson=3.8.8=pypi_0
packaging=23.0=pyhd8ed1ab_0
pandas=1.5.3=py39h417a72b_0
pandas-stubs=1.5.3.230321=pyhd8ed1ab_0
pathtools=0.1.2=py_1
peft=0.3.0.dev0=pypi_0
pillow=9.4.0=py39h6a678d5_0
pip=23.0.1=py39h06a4308_0
platformdirs=3.1.1=pyhd8ed1ab_0
plotly=5.13.1=pyhd8ed1ab_0
pooch=1.7.0=pyha770c72_1
protobuf=3.20.0=pypi_0
psutil=5.9.0=py39h5eee18b_0
py-cpuinfo=9.0.0=pypi_0
pyarrow=11.0.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pydantic=1.10.7=pypi_0
pydub=0.25.1=pypi_0
pyopenssl=23.0.0=py39h06a4308_0
pyparsing=3.0.9=pyhd8ed1ab_0
pyrsistent=0.19.3=pypi_0
pysocks=1.7.1=py39h06a4308_0
python=3.9.16=h7a1cb2a_2
python-dateutil=2.8.2=pyhd8ed1ab_0
python-multipart=0.0.6=pypi_0
python_abi=3.9=2_cp39
pytorch=2.0.0=py3.9_cuda11.7_cudnn8.5.0_0
pytorch-cuda=11.7=h778d358_3
pytorch-mutex=1.0=cuda
pytz=2022.7.1=pyhd8ed1ab_0
pyyaml=6.0=py39hb9d737c_4
readline=8.2=h5eee18b_0
regex=2023.3.22=pypi_0
requests=2.28.1=py39h06a4308_1
requests-oauthlib=1.3.1=pypi_0
responses=0.18.0=pypi_0
rfc3986=1.5.0=pypi_0
rsa=4.9=pypi_0
safetensors=0.3.1=pypi_0
scikit-learn=1.2.1=py39h6a678d5_0
scipy=1.10.0=py39h14f4228_1
semantic-version=2.10.0=pypi_0
sentencepiece=0.1.95=py39hd09550d_0
sentry-sdk=1.17.0=pyhd8ed1ab_0
setproctitle=1.2.2=py39hb9d737c_2
setuptools=65.6.3=py39h06a4308_0
six=1.16.0=pyhd3eb1b0_1
smmap=3.0.5=pyh44b312d_0
sniffio=1.3.0=pypi_0
sqlite=3.41.1=h5eee18b_0
starlette=0.26.1=pypi_0
sympy=1.11.1=py39h06a4308_0
tenacity=8.2.2=pyhd8ed1ab_0
tensorboard=2.12.1=pypi_0
tensorboard-data-server=0.7.0=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
termcolor=2.2.0=pypi_0
textbrewer=0.2.1.post1=pypi_0
threadpoolctl=3.1.0=pyh8a188c0_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.13.2=pypi_0
toolz=0.12.0=pypi_0
torchaudio=2.0.0=py39_cu117
torchtriton=2.0.0=py39
torchvision=0.15.0=py39_cu117
tornado=6.1=py39hb9d737c_3
tqdm=4.65.0=pyhd8ed1ab_1
transformers=4.28.0=pypi_0
types-pytz=2022.7.1.2=pyhd8ed1ab_0
typing-extensions=4.4.0=py39h06a4308_0
typing_extensions=4.4.0=py39h06a4308_0
tzdata=2022g=h04d1e81_0
uc-micro-py=1.0.1=pypi_0
urllib3=1.26.14=py39h06a4308_0
uvicorn=0.21.1=pypi_0
wandb=0.14.0=pyhd8ed1ab_0
wcwidth=0.2.6=pypi_0
websockets=10.4=pypi_0
werkzeug=2.2.3=pypi_0
wheel=0.38.4=py39h06a4308_0
xxhash=3.2.0=pypi_0
xz=5.2.10=h5eee18b_1
yaml=0.2.5=h7f98852_2
yarl=1.7.2=py39hb9d737c_2
zipp=3.15.0=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.2=ha4553b6_0
